+++ src/tests/config/load_config_tests.cpp
@@ -64,8 +64,7 @@ TEST_F(LoadConfigTest, DefaultValueTest) {
   ASSERT_EQ(config_.http_proxy().reuse_kerb(), true);
   ASSERT_EQ(config_.http_proxy().reuse_ntlm(), true);
 
-  ASSERT_EQ(config_.socks_proxy().version(),
-            ssf::network::Socks::Version::kV5);
+  ASSERT_EQ(config_.socks_proxy().version(), ssf::network::Socks::Version::kV5);
   ASSERT_EQ(config_.socks_proxy().host(), "");
   ASSERT_EQ(config_.socks_proxy().port(), "1080");